[Refactor] : Tag 컴포넌트화 & 적용
